Run maintenance tests in-process
