+++ tests/run_e2e_tests.sh
@@ -3,10 +3,10 @@ set -e
 SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 ROOT_DIR="$(cd "$SCRIPT_DIR/.." && pwd)"
 
-E2E_BIN="$SCRIPT_DIR/e2e_tests"
+E2E_BIN="$ROOT_DIR/maint_tests"
 
-gcc -std=c11 -O2 -Wall -Wextra -DUNIT_TEST -I"$ROOT_DIR" \
-    -o "$E2E_BIN" "$SCRIPT_DIR/e2e_tests.c" "$ROOT_DIR/main.c"
+gcc -std=c11 -O2 -Wall -Wextra -DENABLE_INTERNAL_TESTS -I"$ROOT_DIR" \
+    -o "$E2E_BIN" "$ROOT_DIR/main.c"
 
 cd "$ROOT_DIR"
-"$E2E_BIN"
+"$E2E_BIN" --run-e2e-tests
